Внесены изменения в классы лекций
