--- src/main/java/ru/naumen/javakids/model/lectures/Lecture1.java
@@ -1,13 +1,0 @@
-package ru.naumen.javakids.model.lectures;
-
-import ru.naumen.javakids.model.AbstractLecture;
-import ru.naumen.javakids.model.Status;
-
-public class Lecture1 extends AbstractLecture {
-    public Lecture1() {
-        id = 1L;
-        topic = "Знакомство с Java";
-        content = "Содержимое лекции 1";
-        status = Status.NOT_STARTED;
-    }
-}
