--- src/main/java/ru/naumen/javakids/model/lectures/Lecture3.java
@@ -1,13 +1,0 @@
-package ru.naumen.javakids.model.lectures;
-
-import ru.naumen.javakids.model.AbstractLecture;
-import ru.naumen.javakids.model.Status;
-
-public class Lecture3 extends AbstractLecture {
-    public Lecture3() {
-        id = 3L;
-        topic = "Объявление переменных";
-        content = "Содержимое лекции 3";
-        status = Status.NOT_STARTED;
-    }
-}
