+++ src/main/java/ru/naumen/javakids/services/LectureServiceImpl.java
@@ -58,7 +58,7 @@ public class LectureServiceImpl implements LectureService {
 
         if (lectureOp.isPresent()) {
             Lecture updLecture = lectureOp.get();
-            updLecture.setStatus(status);
+          //  updLecture.setStatus(status);
             lectureRepo.save(updLecture);
         }
     }
