+++ src/main/resources/application.properties
@@ -3,4 +3,7 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.jpa.generate-ddl=true
 
+#spring.datasource.initialization-mode=always
+#spring.datasource.data=classpath:data.sql
+
 spring.mustache.expose-request-attributes=true
