+++ src/main/resources/data.sql
@@ -1,0 +1,11 @@
+INSERT INTO public.lectures(
+	id, content, status, topic)
+	VALUES (1, 'Содержимое лекции 1', null, 'Знакомство с Java');
+
+INSERT INTO public.lectures(
+	id, content, status, topic)
+	VALUES (2, 'Содержимое лекции 2', null, 'Структура Java приложения');
+
+INSERT INTO public.lectures(
+	id, content, status, topic)
+	VALUES (3, 'Содержимое лекции 3', null, 'Объявление переменных');
