+++ src/main/resources/templates/layout/header.mustache
@@ -27,7 +27,7 @@
                 {{#user}}
                 <ul class="navbar-nav">
                     <li class="nav-item">
-                        <a class="nav-link" href="/mylectures">Мои лекции</a>
+                        <a class="nav-link" href="/myLectures">Мои лекции</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="/user/{{id}}">Профиль</a>
