+++ src/main/resources/templates/lectures/add.mustache
@@ -1,14 +1,14 @@
 {{> /layout/header}}
 
 <div class="container mt-3">
-    <form action="/createLecture" method="post">
+    <form action="/createLec" method="post">
         <div class="mb-3 mt-3">
             <input type="text" class="form-control" placeholder="Тема лекции" name="topic">
         </div>
         <div class="mb-3">
-            <textarea name="content" rows="10" class="form-control" id="summernote"></textarea>
+            <textarea rows="10" class="form-control" id="summernote" name="content"></textarea>
         </div>
-        <button type="submit" class="btn btn-secondary">Добавить</button>
+        <div><input type="submit" class="btn btn-secondary" value="Добавить"/></div>
     </form>
 </div>
