+++ src/main/resources/templates/lectures/list.mustache
