+++ src/main/resources/templates/lectures/update.mustache
