+++ src/main/resources/templates/user/myLecturesList.mustache
@@ -13,7 +13,7 @@
             <tr>
                 <td>{{id}}</td>
                 <td><a href="/lecture/{{id}}"> {{topic}} </a></td>
-                <td>{{status}}</td>
+
             </tr>
         {{/lectures}}
